+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Translated from CoffeeScript to JavaScript.
 * Added native JavaScript string functions such as `substr()`, `substring()`, `match()`, `indexOf()`, etc.
 * Added `length` property.
+* Renamed `ltrim()` to `trimLeft()` and `rtrim()` to `trimRight()`.
 
 0.2.2 / 2012-09-20
 ------------------